Fix IPA execution issues due to lack of entropy
While running TinyIPA stable/stein images in KVM virtual machines with
Tenks, we noticed that IPA would start but not execute further, which
appeared to be caused by a lack of entropy.

This patch installs haveged, a simple entropy daemon [1], into the
tinyipa ramdisk and starts it on boot, which resolves the issue.

While this problem was reproduced reliably with stable/stein and not
seen with stable/train or stable/rocky images, we will start running
haveged in all versions, for consistency. It adds less than 20 KiB to
the ramdisk.

[1] http://www.issihosts.com/haveged/

Change-Id: I7d47b7828733cc03c1ed6c370c451465a6d492b6
Story: 2006911
Task: 37554