Be a bit more verbose around tinyipa for rescue
Seems we have no real way to grok if there was a rescue setup failure
because we don't log any information to the console about the node
once dhcp has occured.

As such, added an additional line of text to be logged and added
the ip addr command which should save a list of addresses by
interface to the console to assist in CI job troubleshooting.

Change-Id: Ib3f6e24d0687e9da8e551355fe04faf3f86ff582