Test artifact publishing via AFS
I'm requesting this job be temporarily merged for testing publishing
artifacts to an AFS volume instead of pushing it to a static server.
This is part of the efforts to retire our static.openstack.org server
and serve files from our distributed storage [1].

Because the publishing steps happen in the post phase using a secret
value in project-config to get the kerberos ticket for writing to AFS,
we have to commit changes to really test them.

This repository is a good user of the artifact upload process and a
good target for testing the changes on just one repository.

This adds a job ironic-python-agent-build-image-tinyipa-publish-to-afs
which is ultimately parented to the new
publish-openstack-artifacts-with-afs from project-config.  The only
difference is that this will upload the output to AFS, where I will
check that everything copied as expected.  We can then revert this.

Ultimately we will change the publish-openstack-artifacts job in
project-config -- this will be transparent to end-users such as this
repository as we will be switching tarballs.openstack.org to use this
at the same time.

Depends-On: https://review.opendev.org/705159
[1] https://docs.opendev.org/opendev/infra-specs/latest/specs/retire-static.html

Change-Id: Ib2ae8c2d9f59695ce7af5414f73e1f4bcc7ae462