Add ntpdate/ntpclient/chrony
Adds ntpdate to the binary package list for the diskimage-builder
element, and added ntpclient to the tinycore package list and added
a small shim to wrap ntpclient as ntpdate is not a native tinycore
package.

For larger ramdisks, we attempt to use chrony, which superceeds the ntp
package as a combined time server and client.

Change-Id: If99f2df31af8e4d41ad7813d25f0d779abc23f7a
Needed-By: https://review.opendev.org/#/c/707911/