Strip some excess from the centos dib image
Strips postfix gcc make from Centos based builds in
an attempt to reduce the overall sie of the ramdisk as C development
utilties are not needed on a ramdisk and better to not have present
on a machine as it complicates attacks to take over and launch to
additional hosts.

Postfix is also present in the ramdisk and is not needed as the
ramdisk has no use for a mail transport agent.

Also removes the GeoIP data files which are not needed by the ramdisk.

Change-Id: I7627e8a3d4924a672af37a468870ba92604e2e66