Add dosfsutils for EFI to tinyipa
In trying to do a UEFI partition image, we need dosfstools
and our tinycore image doesn't have it present.

Ultimately this raises an exception along the lines of:

 prepare_image, error: Failed to create a file system.
 File system vfat is not supported.:
 ironic_lib.exception.FileSystemNotSupported:
 Failed to create a file system. File system vfat is not
 supported.

Which is rooted in the mkfs.vfat binary not being
present.

 mkfs: failed to execute mkfs.vfat: No such file or directory

I confirmed the tinycore-master image does indeed lack
the dosfstools package. As such adding it to the final
requirements list should enable tinycore to support UEFI
partition image deployments.

Change-Id: I7df395bd2dcc5613a54a21d4aaa14b1e5862e7c3