Do not build iso for ussuri ipa
Since ipa-builder is branchless its code is used directly from master to
test ipa in stable releases.
The iso has been fixed only recently, and since we need to use
ipa-builder version 2.2.0 in stable/ussuri to test, build and publish
the ipa ramdisk, we can't also build an iso in CI or the build job will
break.
As a result we do not buildi an iso for ussuri ironic-python-agent.

Change-Id: I4801e87f5b3c37f113bfc9578c8e2aaa4d4f32f0