Use git hash instead of release to build ipmitool
ipmitool has not had a stable release since a while.
To keep it up to date we clone its repo, pin it to a git commit hash and
compile it from there.
This will also help to avoid patching the code before the compilation as
the changes will be included already in the cloned code.

Change-Id: Ic3c1c0ceace1559d04288b12ab52e343abe56392