Create random character devices so systemd does not crash
Systemd, depending on the version and the configuration,
can attempt to open urandom when a ramdisk boots, before
the udevd is launched to handle the devices. In particular
in FIPS mode, this can result in the ramdisk crashing.

In order to handle this, we pre-create the chracter device
files which can be statically used in the early start sequence
to allow the agent to boot and deploy.

Change-Id: I587e66a0d6b652c4aeca6fc211602d999d4641e8
(cherry picked from commit 94171f104c441f8aea4f8cfadddbb6f0964b09a4)
(cherry picked from commit 7bf70725a6376f0967ca93dba3c6cdab479c59b6)