Remove some extra Centos Packages
Removes webkit and polkit from ramdisks which are not
operationally needed as we don't need to render web pages
in the ramdisk, and we don't need to do system wide policy
enforcement. This does chain them with || true, just to be
on the safe side as we're trying to opportunistically cut
excess cruft in order to save space overall.

Locally, these changes have taken the uncompressed ramdisk
footprint from 780MB to 730MB.

Change-Id: I402a60c61f1b88b98798d9fae52420b1ca256512