Add libaio engine for fio on Centos/Rhel9
fio packaging previously contained everything in one RPM,
but at some point for EL9 builds, this was split into multiple
RPMs where each engine used was put into it's own RPM.

In order for integrated disk benchmarking to work with
introspection, we use python-hardware. Python-hardware uses
fio, with libaio as a default option.

So now, we will attempt to install the other package which
should make users of the feature happier as their ramdisk
images will come with the contents required to do a benchmark
upfront.

Change-Id: I2028ac1cf350d86903a33e4bd91c209887de1e5c