CI: Change image-build to use Ubuntu Jammy
It appears openstack/project-config is still set to focal for image
artifact builds/uploads. After discussing with opendev infra team,
we think we might be crashing the host vm on the focal kernel
leaving us with minimal information to troubleshoot.

The cleanest path forward is to match what we run the check job on
which is Ubuntu Jammy, as a data point before we change the overall
project-config.

Change-Id: I9d6581132569b4a921e8cbc0a1b10c73a6aeba0d
(cherry picked from commit 732852f31fe595d54bb74b0b7f8cad4800b4352e)