Remove USE_PYTHON3 option
The USE_PYTHON3 option has been defaulted to true for
some time, and is no longer used. It was initially added
to enable adoption of python3, we have long moved past
needing to choose between python2 and python3, and
default on python3.

Also, since python2 is no longer supported, this makes
the code cleaner to read.

Change-Id: Ia9e62634c954446b0b47067fcdcd28cac49fbf4b