Add Mellanox Spectrum to firmware to remove list
The Mellanox Spectrum series of ASICs are switch controllers.
The firmware packages which presently represent over 10% of the
size of an average Centos Stream based ironic-python-agent image.
Since this is not something we expect folks to really do,
and since it is for a switch device, it doesn't make sense to
ship these firmware binaries in the stock IPA ramdisk image.

Change-Id: Ic8ecf289f1e03e5ee7f36cc6ded707a25cb138c8
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>