set a maximum systemd journal size
We've noticed a trend with deploys where we seem to use up all of the
free space in the ramdisk, and this could distinctly be systemd.

While we're not logging it, we odn't need infinite logs either because
the default pattern is likely to consume quite a bit more space than we
would like to use.

As such, set a max size.

Related-Bug: 2116135
Change-Id: Iae94278640d79232ec66402c799e4bd6e42336df
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>