Don't fail early if no config drive found
The use of pipefail means that we fail early if we don't
find a matching config drive. This is because grep returns
exit code 1 if no match is found. We can either drop the
pipefail setting, or force grep to return 0. This allows us
to exit gracefully later in the script with code 0.

Closes-Bug: #2120574
Change-Id: I84e348b44b041672ded800a967dae18437f16bba
Signed-off-by: Doug Szumski <doug@stackhpc.com>