Fix firmware cleanup - more.
Some follow-up checks of recent firmware fixes caused me to realize
we've mixxed up the firmware cleanup even more and our defensiveness
on the removal resulted in some confusion.

Ultimately, I guess this is... my second attempt to improve this, but
also attempt to heavily restrict our automatic glob expansion usage.

2025-11-12 18:58:43.824 | + rm -rf /usr/lib/firmware/vicam
2025-11-12 18:58:43.824 | + for item in ${IPA_REMOVE_FIRMWARE//,/ }
2025-11-12 18:58:43.824 | + [[ iwlwifi != '' ]]
2025-11-12 18:58:43.824 | + [[ iwlwifi =~ / ]]
2025-11-12 18:58:43.824 | + rm -rf /usr/lib/firmware/iwlwifi
2025-11-12 18:58:43.824 | + for item in ${IPA_REMOVE_FIRMWARE//,/ }
2025-11-12 18:58:43.824 | + [[ intel/ibt != '' ]]
2025-11-12 18:58:43.824 | + [[ intel/ibt =~ / ]]
2025-11-12 18:58:43.824 | + rm -r -f '/usr/lib/firmware/intel/ibt*'

The net result before this change was the IPA Centos10 build initramfs
files were running around 533 MB, and with this change the file sizes
are down to 464MB.

Change-Id: Ie420a192fbf9ed2fa246c31d799078186a5ca8e7
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>