Revert "Add simple-init by default"
This reverts commit 3f83c9ccb56200c26c21f33d656bfbdbe467ee2d.

Reason for revert: It appears that this change has in-advertently broken the expected behavior, possibly due to ordering or modeling changes inside of glean/simple-init which has been reported but was not discovered in testing due to a lack of review traction on merging the testing over the past year.

https://bugs.launchpad.net/ironic-python-agent-builder/+bug/2133509 may also be related.

Clearly, the next step is to fix testing, verify behavior, and then push forward.

Change-Id: Ib9ab715ea7604b0679c345485a11e770cf8b8d1e
Signed-Off-By: Julia Kreger <juliaashleykreger@gmail.com>