Make multiple integration jobs voting
A risk exists that we just sort of ingore an integration job
failure with an IPA change which might be a legitimate issue.

The key is that we really need to just fix the jobs when they
fail, and not leave jobs in non-voting status for long periods
of times.

So this, in order to help head off merging breaking changes
which might break a non-voting job, elevates those non-voting
jobs to voting.

Change-Id: I77249daf4bebeaead25ea87757de38b272e6ac7a
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>